* don't remove non-existing zfsRoot
git-svn-id: https://shiva.ms.mff.cuni.cz/svn/zlomekfs/branches/zouhar@2974 b158dfce-25fb-0310-9a5b-f29f65ed45b5

+++ zlomekfs/tests/nose-tests/zfs.py
@@ -512,7 +512,8 @@ class ZfsTest(object):
             (Remove zfsRoot)"""
         
         log.debug("teardownClass")
-        shutil.rmtree(cls.zfsRoot)
+        if os.path.isdir(cls.zfsRoot):
+            shutil.rmtree(cls.zfsRoot)
         # self.zfs = None
     
     def setup(self):
